ADD: custom exception handler

+++ src/main/java/com/example/studentmanagementsystem/repository/StudentRepository.java
@@ -10,6 +10,6 @@ import java.util.Optional;
 public interface StudentRepository extends JpaRepository<Student,Long> {
 
       Optional<Student> findByEmail(String email);
-      Optional<Student> findById(int id);
+      Optional<Student> findById(Long id);
 
 }
